refactor(backend): 패키지명, 경로 및 ErrorCode 변경
1. 패키지명 및 경로 domain -> Entity 변경
2. Database 관련 ErrorCode 변경(추후 비즈니스 코드 점검)
