+++ backend/src/main/java/com/kt/onrace/domain/auth/entity/Role.java
@@ -1,13 +1,13 @@
-package com.kt.onrace.domain.auth.domain;
+package com.kt.onrace.domain.auth.entity;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum Gender {
-	MALE("남자"),
-	FEMALE("여자")
+public enum Role {
+	USER("일반 사용자"),
+	ADMIN("관리자")
 
 	;
