+++ backend/src/main/java/com/kt/onrace/domain/auth/repository/UserRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.kt.onrace.domain.auth.domain.User;
+import com.kt.onrace.domain.auth.entity.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
