+++ backend/src/main/java/com/kt/onrace/domain/main/repository/MainRepository.java
@@ -1,7 +1,7 @@
 package com.kt.onrace.domain.main.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.kt.onrace.domain.main.domain.Main;
+import com.kt.onrace.domain.main.entity.Main;
 
 public interface MainRepository extends JpaRepository<Main, Long> {
 }
